refactor: move graph capability into native module (#366)
