+++ web/tests-compose/compose-portal.spec.ts
@@ -1,4 +1,5 @@
 import { expect, type APIRequestContext, type Locator, type Page, test } from "@playwright/test";
+import "./graph-views";
 
 const controlUrl = requiredEnv("DENSE_MEM_CONTROL_URL").replace(/\/$/, "");
 const userUrl = requiredEnv("DENSE_MEM_USER_URL").replace(/\/$/, "");
